# PERF: concat(axis=1) for objects with different indexes