--- a/pandas/core/indexes/api.py
+++ b/pandas/core/indexes/api.py
@@ -227,9 +227,7 @@ def union_indexes(indexes, sort: bool | None = True) -> Index:
 
     def _unique_indices(inds, dtype) -> Index:
         """
-        Convert indexes to lists and concatenate them, removing duplicates.
-
-        The final dtype is inferred.
+        Concatenate indices and remove duplicates.
 
         Parameters
         ----------
@@ -240,6 +238,9 @@ def union_indexes(indexes, sort: bool | None = True) -> Index:
         -------
         Index
         """
+        if all(isinstance(ind, Index) for ind in inds):
+            result = inds[0].append(inds[1:]).unique()
+            return result.astype(dtype, copy=False)
 
         def conv(i):
             if isinstance(i, Index):